emacs: Clean up use-package configuration and fix a Java indentation bug
The cleanups focus on moving as much as possible from `:init` blocks to `:config` blocks.  It also
is more consistent with using function references instead of generic symbols where possible.  It
moves initialization of global modes from `:init` sections to the `elpaca-after-init` hook, which
will speed up startup a bit.

The Java indentation bug arose because the node type names are regexes instead of string literals.
The indentation rules caused block comments to be indented very strangely. The fix was to put
end-of-match anchors in appropriate places, which keeps the fixes to lambda indentation.