emacs: Disable auto-formatting during typing in eglot
This is the LSP feature that lets the LSP server automatically send updates to your buffer as you
type. That is extremely annoying. There are separate configuration options for requested formatting
operations, which are still enabled.

I default overlay type hints to off because they are visually noisy, but leave the feature enabled
so that it can be turned on if desired.