emacs: Use vanilla eglot instead of eglot-java
The eglot-java package was nice because it automated installation of the language server.  However,
I recently ran into problems with it that I don't feel a need to fix.  Now, the bootstrap command
just installs the Java language server and the new configuration uses it directly.

Also:
- Clean up some unused bootstrap code
- Install less with mise
- Avoid upgrading built-in emacs eglot/eldoc/jsonrpc

The mise java packages were a problem in that jdtls was not able to run through them.  Instead, just
rely on native packages.  Removing the python package because the build for 3.12 was not completing.

The eglot upgrade enabled the dape package, but was very expensive time-wise and not essential.